members/views.py
from rest_framework import viewsets
from .serializers import MemberSerializer
from .models import Member

from django.conf import settings
from django.shortcuts import redirect
from django.views.generic.base import TemplateView
from django.views.generic.edit import FormView

from registration import signals
from .forms import RegistrationForm

class MemberViewSet(viewsets.ModelViewSet):
    queryset = Member.objects.all()
    serializer_class = MemberSerializer

