requirements.txt
Django==1.10.7
django-anymail==0.9
django-cors-headers==2.0.2
django-modelcluster==3.0.1
django-registration==2.2
django-taggit==0.22.0
django-tinymce==2.6.0
django-treebeard==4.1.0
djangorestframework==3.6.2
html5lib==0.999999999
oauthlib==2.0.2
olefile==0.44
packaging==16.8
Pillow==4.1.0
PyJWT==1.4.2
pyparsing==2.2.0
python-openid==2.2.5
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
Unidecode==0.4.20
wagtail==1.9
webencodings==0.5
Willow==0.4
